organizing and pruning
