+++ pkg/tracing/otlp/otlp.go
@@ -5,7 +5,6 @@ package otlp
 
 import (
 	"context"
-	"time"
 
 	"github.com/go-kit/log"
 	"github.com/pkg/errors"
@@ -20,27 +19,6 @@ import (
 	_ "google.golang.org/grpc/encoding/gzip"
 	"gopkg.in/yaml.v2"
 )
-
-type retryConfig struct {
-	RetryEnabled         bool          `yaml:"retry_enabled"`
-	RetryInitialInterval time.Duration `yaml:"retry_initial_interval"`
-	RetryMaxInterval     time.Duration `yaml:"retry_max_interval"`
-	RetryMaxElapsedTime  time.Duration `yaml:"retry_max_elapsed_time"`
-}
-
-type Config struct {
-	ClientType         string            `yaml:"client_type"`
-	ReconnectionPeriod time.Duration     `yaml:"reconnection_period"`
-	Compression        string            `yaml:"compression"`
-	Insecure           bool              `yaml:"insecure"`
-	Endpoint           string            `yaml:"endpoint"`
-	URLPath            string            `yaml:"url_path"`
-	Timeout            time.Duration     `yaml:"timeout"`
-	RetryConfig        retryConfig       `yaml:"retry_config"`
-	Headers            map[string]string `yaml:"headers"`
-}
-
-// add TLS config
 
 var (
 	TracingClientGRPC string = "grpc"
@@ -90,108 +68,4 @@ func NewTracerProvider(ctx context.Context, logger log.Logger, conf []byte) (*tr
 	return tp, nil
 }
 
-func traceGRPCOptions(config Config) []otlptracegrpc.Option {
-	var options []otlptracegrpc.Option
-	if config.Endpoint != "" {
-		options = append(options, otlptracegrpc.WithEndpoint(config.Endpoint))
-	}
 
-	if config.Insecure {
-		options = append(options, otlptracegrpc.WithInsecure())
-	}
-
-	if config.ReconnectionPeriod != 0 {
-		options = append(options, otlptracegrpc.WithReconnectionPeriod(config.ReconnectionPeriod))
-	}
-
-	if config.Timeout != 0 {
-		options = append(options, otlptracegrpc.WithTimeout(config.Timeout))
-	}
-
-	if config.Compression != "" {
-		if config.Compression == "gzip" {
-			options = append(options, otlptracegrpc.WithCompressor(config.Compression))
-		}
-	}
-
-	if config.RetryConfig.RetryEnabled {
-		options = append(options, otlptracegrpc.WithRetry(createGRPCRetryConfig(config)))
-	}
-
-	if config.Headers != nil {
-		options = append(options, otlptracegrpc.WithHeaders(config.Headers))
-	}
-
-	return options
-}
-
-func traceHTTPOptions(config Config) []otlptracehttp.Option {
-	var options []otlptracehttp.Option
-	if config.Endpoint != "" {
-		options = append(options, otlptracehttp.WithEndpoint(config.Endpoint))
-	}
-
-	if config.Insecure {
-		options = append(options, otlptracehttp.WithInsecure())
-	}
-
-	if config.URLPath != "" {
-		options = append(options, otlptracehttp.WithURLPath(config.URLPath))
-	}
-
-	if config.Compression != "" {
-		if config.Compression == "gzip" {
-			options = append(options, otlptracehttp.WithCompression(otlptracehttp.GzipCompression))
-		}
-	}
-
-	if config.Timeout != 0 {
-		options = append(options, otlptracehttp.WithTimeout(config.Timeout))
-	}
-
-	if config.RetryConfig.RetryEnabled {
-		options = append(options, otlptracehttp.WithRetry(createHTTPRetryConfig(config)))
-	}
-
-	if config.Headers != nil {
-		options = append(options, otlptracehttp.WithHeaders(config.Headers))
-	}
-
-	return options
-}
-
-func createHTTPRetryConfig(config Config) otlptracehttp.RetryConfig {
-
-	var retryConfig otlptracehttp.RetryConfig
-	if config.RetryConfig.RetryInitialInterval != 0 {
-		retryConfig.InitialInterval = config.RetryConfig.RetryInitialInterval
-	}
-
-	if config.RetryConfig.RetryMaxInterval != 0 {
-		retryConfig.MaxInterval = config.RetryConfig.RetryMaxInterval
-	}
-
-	if config.RetryConfig.RetryMaxElapsedTime != 0 {
-		retryConfig.MaxElapsedTime = config.RetryConfig.RetryMaxElapsedTime
-	}
-
-	return retryConfig
-}
-
-func createGRPCRetryConfig(config Config) otlptracegrpc.RetryConfig {
-
-	var retryConfig otlptracegrpc.RetryConfig
-	if config.RetryConfig.RetryInitialInterval != 0 {
-		retryConfig.InitialInterval = config.RetryConfig.RetryInitialInterval
-	}
-
-	if config.RetryConfig.RetryMaxInterval != 0 {
-		retryConfig.MaxInterval = config.RetryConfig.RetryMaxInterval
-	}
-
-	if config.RetryConfig.RetryMaxElapsedTime != 0 {
-		retryConfig.MaxElapsedTime = config.RetryConfig.RetryMaxElapsedTime
-	}
-
-	return retryConfig
-}
